third_party/blink/web_tests/platform/generic/fast/dom/iframe-inner-size-scaling-expected.txt

This tests that innerWidth/innerHeight on an frame window returns the size of the frame itself in CSS pixels, regardless of page scale.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS frame.contentWindow.innerWidth is non-zero.
PASS frame.contentWindow.innerHeight is non-zero.
PASS frame.contentWindow.innerWidth is originalWidth
PASS frame.contentWindow.innerHeight is originalHeight
PASS successfullyParsed is true

TEST COMPLETE
